tweak prediction algorithm. But it still does not work.

+++ src/main/scala/wh/models/domain.scala
@@ -4,6 +4,6 @@ case class User(name: String)
 
 case class Location(name: String)
 
-case class Rating(user: User, location: Location, rate: Int) {
+case class Rating(user: User, location: Location, rate: Double) {
   require (rate >= 1 && rate <= 5)
 }
